+++ tradingstrategy/exchange.py
@@ -187,9 +187,19 @@ class ExchangeUniverse:
                 return xchg
         return None
 
+    def get_by_chain_and_factory(self, chain_id: ChainId, factory_address: str) -> Optional[Exchange]:
+        """Get the exchange implementation on a specific chain.
 
+        :param chain_id: Blockchain this exchange is on
 
-
+        :param factory_address: The smart contract address of the exchange factory
+        """
+        assert isinstance(chain_id, ChainId)
+        factory_address = factory_address.lower()
+        for xchg in self.exchanges.values():
+            if xchg.address.lower() == factory_address:
+                return xchg
+        return None